+++ packages/select/src/select.tsx
@@ -105,6 +105,10 @@ export default defineComponent({
     filterOption: { type: Function }, // 配置当前options的过滤规则
     searchWithPinyin: PropTypes.bool.def(true), // 拼音搜索
     highlightKeyword: PropTypes.bool.def(false), // 搜索高亮
+    trigger: {
+      type: String as PropType<'default' | 'manual'>,
+      default: 'default',
+    }, // content显示和隐藏方式
   },
   emits: [
     'update:modelValue',
@@ -155,6 +159,7 @@ export default defineComponent({
       searchWithPinyin,
       highlightKeyword,
       disableFocusBehavior,
+      trigger,
     } = toRefs(props);
 
     const virtualRenderRef = ref(null);
@@ -203,7 +208,7 @@ export default defineComponent({
     const searchRef = ref<HTMLElement>();
     const selectTagInputRef = ref<SelectTagInputType>();
     const popoverRef = ref();
-    const optionsMap = ref<Map<any, OptionInstanceType>>(new Map());
+    const optionsMap = ref<Map<PropertyKey, OptionInstanceType>>(new Map());
     const options = computed(() =>
       [...optionsMap.value.values()].sort((cur, next) => {
         return cur.order - next.order;
@@ -211,12 +216,13 @@ export default defineComponent({
     );
     const groupsMap = ref<Map<string, GroupInstanceType>>(new Map());
     const selected = ref<ISelected[]>([]);
-    const selectedMap = computed<Record<string, string>>(() =>
+    const selectedMap = computed<Record<PropertyKey, string>>(() =>
       selected.value.reduce((pre, item) => {
         pre[item.value] = item.label;
         return pre;
       }, {}),
     );
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const activeOptionValue = ref<any>(); // 当前悬浮的option
     const listMap = computed(() =>
       list.value.reduce((pre, item) => {
@@ -364,6 +370,14 @@ export default defineComponent({
       { popoverMinWidth: popoverMinWidth.value },
       triggerRef,
     );
+    const handleHidePopover = () => {
+      if (trigger.value === 'manual') return;
+      hidePopover();
+    };
+    const handleShowPopover = () => {
+      if (trigger.value === 'manual') return;
+      showPopover();
+    };
     // 输入框是否可以输入内容
     const isInput = computed(
       () => ((filterable.value && inputSearch.value) || allowCreate.value) && isPopoverShow.value,
@@ -405,6 +419,7 @@ export default defineComponent({
       }
     };
     // 默认搜索方法
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const defaultSearchMethod = (searchValue: string, optionName: string, filterData: Record<string, any> = {}) => {
       if (hasFilterOptionFunc.value) {
         // 是否配置了单个options过滤
@@ -463,7 +478,7 @@ export default defineComponent({
     };
     // 派发toggle事件
     const handleTogglePopover = () => {
-      if (isDisabled.value) return;
+      if (isDisabled.value || trigger.value === 'manual') return;
       handleFocus();
       togglePopover();
     };
@@ -501,7 +516,7 @@ export default defineComponent({
       } else {
         selected.value = [{ value, label: value }];
         emitChange(value);
-        hidePopover();
+        handleHidePopover();
       }
       customOptionName.value = '';
     };
@@ -541,7 +556,7 @@ export default defineComponent({
         ];
         emitChange(option.optionID);
         emit('select', option.optionID);
-        hidePopover();
+        handleHidePopover();
         handleBlur();
       }
     };
@@ -576,7 +591,7 @@ export default defineComponent({
       selected.value = [];
       emitChange(multiple.value ? [] : '');
       emit('clear', multiple.value ? [] : '');
-      hidePopover();
+      handleHidePopover();
     };
     const handleSelectedAllOptionMouseEnter = () => {
       activeOptionValue.value = '';
@@ -641,7 +656,7 @@ export default defineComponent({
       }
     };
     // options存在 > 上一次选择的label > 当前值
-    const handleGetLabelByValue = (value: string) => {
+    const handleGetLabelByValue = (value: PropertyKey) => {
       // 处理options value为对象类型，引用类型变更后，回显不对问题
       let tmpValue = value;
       if (typeof tmpValue === 'object') {
@@ -681,6 +696,13 @@ export default defineComponent({
           selected.value = [];
         }
       }
+    };
+    // 手动设置selected值
+    const setSelected = (data: Array<object>) => {
+      selected.value = data.map(item => ({
+        label: item[displayKey.value],
+        value: item[idKey.value],
+      }));
     };
     // 处理键盘事件
     const handleDocumentKeydown = (e: KeyboardEvent) => {
@@ -735,7 +757,7 @@ export default defineComponent({
     const handleClickOutside = ({ event }) => {
       const { target } = event;
       if (triggerRef.value?.contains(target) || triggerRef.value === target) return;
-      hidePopover();
+      handleHidePopover();
       handleBlur();
     };
 
@@ -746,7 +768,7 @@ export default defineComponent({
         selected,
         activeOptionValue,
         showSelectedIcon,
-        selectedStyle: selectedStyle as any, // todo 类型推断
+        selectedStyle,
         curSearchValue,
         highlightKeyword,
         register,
@@ -761,7 +783,7 @@ export default defineComponent({
     onMounted(() => {
       handleSetSelectedData();
       setTimeout(() => {
-        showOnInit.value && showPopover();
+        showOnInit.value && handleShowPopover();
         autoFocus.value && focusInput();
       });
     });
@@ -832,6 +854,7 @@ export default defineComponent({
       isEnableVirtualRender,
       preloadItemCount,
       virtualRenderRef,
+      setSelected,
     };
   },
   render() {
@@ -931,6 +954,7 @@ export default defineComponent({
             v-slots={{
               prefix: renderPrefix(),
               default: this.$slots?.tag && (() => this.$slots?.tag({ selected: this.selected })),
+              tagRender: this.$slots?.tagRender && ((item: ISelected) => this.$slots?.tagRender(item)),
               suffix: () => suffixIcon(),
             }}
             behavior={this.behavior}
@@ -994,6 +1018,7 @@ export default defineComponent({
         >
           {{
             default: ({ data }) => {
+              // 兼容以前slots
               const optionRender = this.$slots?.optionRender || this.$slots?.virtualScrollRender;
               return data.map(item => (
                 <Option
@@ -1007,14 +1032,18 @@ export default defineComponent({
           }}
         </VirtualRender>
       ) : (
-        this.filterList.map(item => (
-          <Option
-            id={item[this.idKey]}
-            key={item[this.idKey]}
-            v-slots={this.$slots?.optionRender ? { default: () => this.$slots?.optionRender?.({ item }) } : null}
-            name={item[this.displayKey]}
-          />
-        ))
+        this.filterList.map(item => {
+          // 兼容以前slots
+          const optionRender = this.$slots?.optionRender || this.$slots?.virtualScrollRender;
+          return (
+            <Option
+              id={item[this.idKey]}
+              key={item[this.idKey]}
+              v-slots={typeof optionRender === 'function' ? { default: () => optionRender({ item }) } : null}
+              name={item[this.displayKey]}
+            />
+          );
+        })
       );
     };
     // 渲染内容