+++ site/views/table/config.ts
@@ -99,7 +99,7 @@ export default [
         name: 'row-height',
         type: 'Number|Function',
         default: 'LINE_HEIGHT',
-        desc: '行高，可以为固定数值类型, 可以是函数，返回当前行的高度，返回值为数值类型',
+        desc: '行高，可以为固定数值类型, 可以是函数，返回当前行的高度，返回值为数值类型; 如果是回调函数，请注意参数：args: { index: number; type: tbody|virtual; row?: any; items?: number[] }; type = tbody 说明是table tr渲染时请求计算tr行高，此时 row 为当前行数据； 如果 type = virtual 说明启用了虚拟渲染，此时是计算渲染行高，此时这里没有row属性， 可以使用 index 自行获取 row',
         optional: [],
       },
       {
@@ -331,6 +331,34 @@ export default [
         }`,
         ],
       },
+      {
+        name: 'empty-cell-text',
+        type: 'String | ({ row, column, index }) => string | jsx.Element | VNode',
+        default: 'true',
+        desc: '单元格数据为空展示',
+        optional: [],
+      },
+      {
+        name: 'is-empty-cell',
+        type: '["", undefined, null] | ({ cellText, row, column }) => boolean',
+        default: '["", undefined, null]',
+        desc: `  /**
+         * 判定单元格是否为空
+         * 支持数组：判定条件为 arr.some(item => item === cellText)
+         * 支持函数回调：返回结果为 true | false， ({ cellText, row, column }) => boolean
+         */`,
+        optional: [],
+      },
+      {
+        name: 'append-last-row',
+        type: 'Option',
+        default: '{ type: "default", cellRender: undefined }',
+        desc: '表格尾部追加的行配置, 这里的 type = default时，会优选渲染 cellRender 函数，如果没有配置，则会渲染插槽 #appendLastRow； 如果要实现统计功能，请配置 type: "summary"，每一列的渲染统计显示通过回调函数 cellRender 返回；',
+        optional: [
+          '{ type: "default", cellRender?: (column: Column, index: number) => VNode | number | string }',
+          '{ type: "summary", cellRender?: (column: Column, index: number) => VNode | number | string }',
+        ],
+      },
     ],
   },
   {
@@ -433,6 +461,13 @@ export default [
         desc: '嵌套实现多表头，如果是多表头分组，可以只设置 label 属性',
         optional: [],
       },
+      {
+        name: 'acrossPage',
+        type: 'Boolean',
+        default: 'undefined',
+        desc: '需要配合`type=selection`使用，是否启用跨页全选，如果设置为true，表头会渲染跨页全选功能和相应插槽',
+        optional: [],
+      },
     ],
   },
   {
@@ -527,7 +562,13 @@ export default [
         optional: ['def1', 'def2'],
       },
       { name: 'isShow', type: 'Boolean', default: 'true', desc: '是否显示Head', optional: [] },
-      { name: 'cellFn', type: 'Function', default: 'undefined', desc: '自定义当前列渲染函数', optional: [] },
+      {
+        name: 'cellFn',
+        type: '({ index, column }) => string | VNode | JSX.Element',
+        default: 'undefined',
+        desc: '自定义当前列渲染函数',
+        optional: [],
+      },
     ],
   },
   {
@@ -749,6 +790,11 @@ export default [
       {
         name: '#fixedBottom',
         desc: '底部加载插槽,此插槽内容会一直固定在底部, 可以结合 props.fixedBottom 进行详细配置',
+        params: '',
+      },
+      {
+        name: '#appendLastRow',
+        desc: '追加到最后一行插槽，区别于 #fixedBottom, 此插槽内容会追加最后一行数据之后，可以滚动，而fixedBottom内容固定底部（如果是虚拟滚动，这里暂时不支持）；注意：如果设置 append-last-row type=default时，会优先渲染 append-last-row 中配置的cellRender函数；',
         params: '',
       },
       { name: '#setting', desc: '表格设置中间自定义插槽', params: '' },