+++ site/views/tree/config.ts
@@ -47,6 +47,13 @@ export default [
       { name: 'indent', type: 'Number', default: '16', desc: '相邻级节点间的水平缩进，单位为像素', optional: [] },
       { name: 'line-height', type: 'Number', default: '32', desc: '设置行高', optional: [] },
       {
+        name: 'height',
+        type: 'Number',
+        default: '-',
+        desc: '设置树形组件高度；在设置 virtualRender=true时，请指定高度，避免组件自动计算高度导致多次渲染',
+        optional: [],
+      },
+      {
         name: 'level-line',
         type: 'Boolean|String|Function',
         default: 'false',
@@ -294,7 +301,15 @@ export default [
         desc: '设置指定节点是否展开',
         params: '(item: any, isOpen = null, e: MouseEvent = null, fireEmit = true)',
       },
-      { name: 'setSelect', desc: '设置指定节点是否选中', params: '(item: any, selected = true, autoOpen = true)' },
+      {
+        name: 'setSelect',
+        desc: `* 设置节点选中状态
+           * @param nodes 选中节点，可以是多个
+           * @param selected 是否选中 default：true
+           * @param autoOpen 是否自动展开所有父级节点 default：true
+           * @param triggerEvent 是否触发抛出事件 false`,
+        params: '(item: any, selected = true, autoOpen = true, triggerEvent=false)',
+      },
       { name: 'asyncNodeClick', desc: '异步请求触发点击节点', params: '（item）' },
       { name: 'getData', desc: '获取当前树配置数据（经过内部处理的数据）', params: '（）' },
       { name: 'getParentNode', desc: '获取指定节点的父级节点', params: 'node: 当前节点' },